+++ src/ssd1351/ssd1351.hpp
@@ -84,7 +84,7 @@ namespace upm {
 /**
  * @library ssd1351
  * @sensor ssd1351
- * @comname SSD1351 OLED
+ * @comname SPI-based OLED Display
  * @type display
  * @man adafruit
  * @web http://www.adafruit.com/products/1431
@@ -112,7 +112,7 @@ class SSD1351 : public GFX{
          * @param dc Data/command pin
          * @param rst Reset pin
          */
-        SSD1351 (uint8_t oc, uint8_t dc, uint8_t rst);
+        SSD1351 (int oc, int dc, int rst);
 
         /**
          * SSD1351 object destructor
